Rename UnauthenticatedApp to NotAuhenticatedApp
