+++ src/components/NotAuthenticatedApp/NotAuthenticatedApp.tsx
