+++ src/components/NotAuthenticatedApp/index.ts
